Rename auth to genericauth
