+++ handlers/auth_test.go
@@ -8,7 +8,6 @@ import (
 	"testing"
 
 	"github.com/mtlynch/screenjournal/v2"
-	"github.com/mtlynch/screenjournal/v2/auth"
 	"github.com/mtlynch/screenjournal/v2/auth2"
 	"github.com/mtlynch/screenjournal/v2/handlers"
 	"github.com/mtlynch/screenjournal/v2/handlers/sessions"
@@ -140,7 +139,7 @@ func TestAuthPost(t *testing.T) {
 }
 
 func mustCreatePasswordHash(plaintext string) screenjournal.PasswordHash {
-	h, err := auth.NewPasswordHash(plaintext)
+	h, err := auth2.NewPasswordHash(plaintext)
 	if err != nil {
 		panic(err)
 	}
